+++ app/src/main/kotlin/io/github/wykopmobilny/ui/modules/profile/ProfileActivityV2.kt
@@ -10,6 +10,15 @@ import io.github.wykopmobilny.utils.viewBinding
 
 internal class ProfileActivityV2 : ThemableActivity() {
 
+    companion object {
+        const val EXTRA_USER_ID = "EXTRA_USERNAME"
+
+        fun createIntent(context: Context, userId: String) =
+            Intent(context, ProfileActivityV2::class.java).apply {
+                putExtra(EXTRA_USER_ID, userId)
+            }
+    }
+
     private val binding by viewBinding(ActivityContainerBinding::inflate)
 
     private val userId
@@ -23,14 +32,5 @@ internal class ProfileActivityV2 : ThemableActivity() {
                 .replace(binding.fragmentContainer.id, profileMainFragment(userId))
                 .commit()
         }
-    }
-
-    companion object {
-        const val EXTRA_USER_ID = "EXTRA_USERNAME"
-
-        fun createIntent(context: Context, userId: String) =
-            Intent(context, ProfileActivityV2::class.java).apply {
-                putExtra(EXTRA_USER_ID, userId)
-            }
     }
 }
